# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Service Swagger",
        "version": "1.0.0"
    },
    "paths": {
        "/api/reward": {
            "post": {
                "tags": [
                    "Create reward"
                ],
                "summary": "Create reward",
                "description": "Reward Creation Endpoint",
                "operationId": "createReward",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "details"
                                ],
                                "properties": {
                                    "details": {
                                        "type": "string",
                                        "example": "{\"title\":\"test reward\"}"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reward created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            }
        },
        "/api/reward/{id}": {
            "get": {
                "tags": [
                    "Retrieve reward"
                ],
                "summary": "Retrieve reward",
                "description": "Reward Retrieve Endpoint",
                "operationId": "showReward",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reward id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reward retrieved",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Update reward"
                ],
                "summary": "Update reward",
                "description": "Reward Update Endpoint",
                "operationId": "updateReward",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reward id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "details"
                                ],
                                "properties": {
                                    "details": {
                                        "type": "string",
                                        "example": "{\"title\":\"test reward\"}"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reward updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Delete reward"
                ],
                "summary": "Delete reward",
                "description": "Reward Delete Endpoint",
                "operationId": "deleteReward",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Reward id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reward deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            }
        },
        "/api/reward/{id}/attach_to_user": {
            "post": {
                "tags": [
                    "Attach user to reward"
                ],
                "summary": "Attach user to reward",
                "description": "Attach user to reward Endpoint",
                "operationId": "attachReward",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "uid"
                                ],
                                "properties": {
                                    "uid": {
                                        "type": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reward attached Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Record not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "api_key_security": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer *token*)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}